Add files via upload
Code to run initially to store student data on RFID Cards/Tags. First reset the nodeMCU. Then start to store the data. While storing the data of each student in the card, use the serial monitor. Make sure to hold the card very near to the RFID Scanner until you feed the whole data of the student in the card. After storing data in first card then reset the the NodeMCU to store data on next card.